requirements.txt
flask
waitress
pyyaml
flask_httpauth
